+++ cyc-admin_panel/src/components/configTiers.component.js
@@ -28,15 +28,17 @@ export default class ConfigTiers extends Component {
     }
 
     remove(id) {
-        console.log("Deleting");
-        DataService.remove(id)
-            .then(response => {
-                console.log(response.data);
-                window.location.reload(false);
-            })
-            .catch(e => {
-                console.log(e);
-            });
+        if (window.confirm(`Confirm deletion - ${id}`)) {
+            console.log("Deleting");
+            DataService.remove(id)
+                .then(response => {
+                    console.log(response.data);
+                    window.location.reload(false);
+                })
+                .catch(e => {
+                    console.log(e);
+                });
+        }
     }
     render() {
         const { DataisLoaded, items } = this.state;
