+++ cyc-admin_panel/src/components/configTracks.component.js
@@ -25,15 +25,17 @@ export default class ConfigTracks extends Component {
   }
 
   remove(id) {
-    console.log("Deleting");
-    DataService.remove(id)
-      .then((response) => {
-        console.log(response.data);
-        window.location.reload(false);
-      })
-      .catch((e) => {
-        console.log(e);
-      });
+    if (window.confirm(`Confirm deletion - ${id}`)) {
+      console.log("Deleting");
+      DataService.remove(id)
+        .then((response) => {
+          console.log(response.data);
+          window.location.reload(false);
+        })
+        .catch((e) => {
+          console.log(e);
+        });
+    }
   }
 
   render() {
